Temporarily isolate spray usages in separate package
